Organize configuration-related internal classes
