+++ src/main/java/com/greghaskins/spectrum/internal/Child.java
@@ -1,6 +1,8 @@
 package com.greghaskins.spectrum.internal;
 
 import com.greghaskins.spectrum.Block;
+import com.greghaskins.spectrum.internal.configuration.ConfiguredBlock;
+import com.greghaskins.spectrum.internal.configuration.TaggingFilterCriteria;
 
 import org.junit.runner.Description;
 import org.junit.runner.notification.RunNotifier;
