+++ src/main/java/com/greghaskins/spectrum/internal/configuration/BlockTagging.java
@@ -1,4 +1,6 @@
-package com.greghaskins.spectrum.internal;
+package com.greghaskins.spectrum.internal.configuration;
+
+import com.greghaskins.spectrum.internal.Child;
 
 import java.util.Arrays;
 import java.util.HashSet;
