+++ src/main/java/com/greghaskins/spectrum/internal/configuration/ConfiguredBlock.java
@@ -1,4 +1,4 @@
-package com.greghaskins.spectrum.internal;
+package com.greghaskins.spectrum.internal.configuration;
 
 import com.greghaskins.spectrum.Block;
